+++ packages/date-picker/src/DatePicker/DatePickerInput/index.ts
@@ -1,0 +1,2 @@
+export { DatePickerInput } from './DatePickerInput';
+export { DatePickerInputProps } from './DatePickerInput.types';