+++ packages/date-picker/src/shared/utils/isEverySegmentFilled/isEverySegmentFilled.ts
@@ -1,0 +1,7 @@
+import { isNotZeroLike } from '@leafygreen-ui/lib';
+
+import { DateSegmentsState } from '../../types';
+
+export const isEverySegmentFilled = (segments: DateSegmentsState) => {
+  return Object.values(segments).every(isNotZeroLike);
+};