+++ packages/marketing-modal/src/index.ts
@@ -1,3 +1,6 @@
 export { default } from './MarketingModal/MarketingModal';
-export type { MarketingModalProps } from './MarketingModal/types';
-export { BlobPosition, GraphicStyle } from './MarketingModal/types';
+export type { MarketingModalProps } from './MarketingModal/MarketingModal.types';
+export {
+  BlobPosition,
+  GraphicStyle,
+} from './MarketingModal/MarketingModal.types';