+++ packages/popover/src/Popover/Popover.spec.tsx
@@ -209,13 +209,13 @@ describe('packages/popover', () => {
 
   // eslint-disable-next-line jest/no-disabled-tests
   describe.skip('types', () => {
-    test('default', () => {
-      <Popover>Popover Content</Popover>;
-    });
-
     test('requires children', () => {
       // @ts-expect-error
       <Popover></Popover>;
+    });
+
+    test('Requires only children', () => {
+      <Popover>Popover Content</Popover>;
     });
 
     test('does not allow specifying "portalClassName", when "usePortal" is false', () => {
@@ -223,6 +223,15 @@ describe('packages/popover', () => {
       <Popover active usePortal={false} portalClassName="test-classname">
         Popover Content
       </Popover>;
+    });
+
+    test('accepts transition lifecycle props', () => {
+      <Popover onEnter={() => {}}>test</Popover>;
+      <Popover onEntering={() => {}}>test</Popover>;
+      <Popover onEntered={() => {}}>test</Popover>;
+      <Popover onExit={() => {}}>test</Popover>;
+      <Popover onExiting={() => {}}>test</Popover>;
+      <Popover onExited={() => {}}>test</Popover>;
     });
 
     test('accepts `div` props', () => {